requirements.txt
dspy-ai
streamlit
pandas